Dilpreet | Test | White label Fixes
